feat: implement AmbulanciaSimulationWatcher and abstract SimulationWatcher class; remove AmbulanciaActor
